Added the Lua function playSound to play sounds. Unfortunately, the function still does not use TaskIds, so the usual block-until-finished functionality currently just blocks indefinitely. This should be addressed in the next iteration. Added the thunder sound effect for testing purposes.
Changed the DialogueController to grant the ability to embed scripts in lines of dialogue. DialogueController::Line has been changed into a full-fledged class that parses and indexes embedded scripts, and extracts them for use by the DialogueController.
The DialogueController now keeps a pointer to the ScriptEngine, which it calls on to execute embedded scripts.
Script objects have been changed to load scripts after construction. The constructor just sets the luaStack to NULL (to indicate an uninitialized script). If the Script is run prior to a Script::load* call, then an exception is thrown. The ScriptEngine calls either loadFile to load in an entire script file, or loadString to load in a string of script code for the Script to run. This load now occurs in ScriptEngine::runScript and ScriptEngine::runScriptString after construction.
Added a couple of new items to the to-do list and re-organized it a little bit.

git-svn-id: http://robothaus.org/repos/ed@36 5033be24-7f03-4d61-c009-87962669689c