The main feature of this change is the ability to make playSound script yield until a sound has finished playing. Changed the chapter1 script to take advantage of this new functionality as an exercise.
Created a new Task object to encapsulate task construction and task completion. Task objects are used to signal the Scheduler, encapsulating the Scheduler pointer so that other classes don't need to keep track of the Scheduler. Most of the code now passes around Task objects instead of just TaskIds, and Task::signal is called on task completion. Task objects *self-destruct* after signal is called, since the task is no longer necessary after it is finished. The class has a private destructor in order to prevent creation on the stack.

Sound has, obviously, been fixed up so get and signal Tasks when a Sound has finished playing.

FYI, I love this project.

git-svn-id: http://robothaus.org/repos/ed@38 5033be24-7f03-4d61-c009-87962669689c