Created a stub 'addNPC' function to be used when the NPC system design gets a bit more fleshed out.
Introduced a "delay" function to add explicit script delays. Created a Timer Thread that counts down and signals a task on completion in order to create delays.

In an effort to clean up the global src/ folder, moved some of the coroutine based mechanisms to a new folder: Coroutines. Coroutines/ now includes the Scheduler class, the Task class, the Thread class, and the new Timer class. These classes have been removed from the global space as needed. The project file has been updated to include Coroutines/ as necessary, of course.

git-svn-id: http://robothaus.org/repos/ed@40 5033be24-7f03-4d61-c009-87962669689c