A lot of the files touched here were just documentation additions/cleanups: Task.h, Timer.h, GameState.h, Singleton.h, NPC.h, ResourceKey.h, Resource.h.
Made Thread::~Thread() virtual because I was being an idiot when I made it and forgot that polymorphic classes ALWAYS need virtual destructors or you get some seriously messed up memory corruption.

Made Script an abstract class, and created the FileScript and StringScript classes. Moved initializations into the constructors because the loadFile/loadString business was funny and stupid. FileScript is now used for any script that loads from file, while StringScript is used for any lines of Lua code that are embedded in a string.
Changed ScriptEngine and NPC to use FileScript and StringScript as appropriate.

git-svn-id: http://robothaus.org/repos/ed@43 5033be24-7f03-4d61-c009-87962669689c