Oh my gosh, NPCs are finally in the mix! Granted they don't show up on screen yet, but they still initialize and act independently of each other just as needed. A LOT of changes were required in order to get this working.
- Added two test NPC scripts in Sereia.outside that idle at different speeds.

- Changed Resource::getName to Resource::getResourceName to reduce the ambiguity stemming from use of "getName" as a function name.

- Added a "deletedThreads" list to Scheduler, because deleting Threads in the middle of the run caused weird behaviour (namely that new Threads would be created in the newly freed memory and thus would be wrongly placed in the finished list). The new list defers deletion of ended Threads to the end of a Scheduler frame.

- Added two new debug flags for NPCs and the Scheduler, since the Script Engine debug flag was being overused.

- Script objects now have script names, which need to be supplied to the Script constructor by child classes.
- Script objects now have a running flag to indicate if a Script is in the middle of a run or not. This flag can be queried publicly, but should not have to be publicly read if NPCThread is properly merged with NPCScript.

- When waiting, Threads no longer block on a Task signalled by a Timer. Instead, they just do a join on the Timer thread, which finishes when time is up.

- ScriptFactory now creates and returns NPCScript objects on request
- Documentation update on ScriptFactory

- TileEngine is now solely responsible for the creation of NPCs, instead of receiving NPCs from the ScriptEngine.
- TileEngine now contains a name-to-NPC mapping of all active NPCs on the map.

- ScriptEngine::getNPCScript was added to query the ScriptFactory to return a new NPC script.

How NPCs work:

NPCs contain NPCThreads, which are separate entities that managed the thread/script state of an NPC's AI behaviour. An NPCThread contains (for now) and runs an NPCScript based on whether or not the script is currently in the middle of a run or if the NPC itself currently has instructions to work on. If the script is not running and the NPC is idle, the NPCThread runs the NPC's idle function (if it exists).
The NPC itself is updated by the TileEngine, and the NPCThread is given time to run by the Scheduler as necessary, and updates the NPC's instruction queue via the script instructions (which don't exist yet, but will).

Next steps include merging NPCThread and NPCScript ASAP, and then creating Sprite objects for NPC sprites, as well as movement functions to move the NPCs around independently.

git-svn-id: http://robothaus.org/repos/ed@47 5033be24-7f03-4d61-c009-87962669689c