This change is fairly far-reaching, and concerns itself with the game's lack of graceful failure when a resource fails to load. Prior to this change, an inability to load a music or sound file would result in a segmentation fault, which is terrible. So here is the gist of the changes:
---BETTER DEBUGGING---
- Added a "DEBUG_MODE" macro to easily toggle debug utilities (currently just output and pausing)
- Added a "DEBUG_PAUSE" macro to pause all processing to make reading and interpreting debug output easier
- Added Debug output and pausing to main.cpp, as well as a corresponding debug flag in DebugUtils.h

---STOP THE MUSIC!---
- Removed the MUSIC_OFF macro from MainMenu.cpp since DebugUtils already has it

---FAST TALKING---
- Doubled the dialogue speed for quicker testing

---BASIC RESOURCE CHANGES---
- Removed ResourceException (and all its inclusions) since it was only being caught in one specific spot anyway, and didn't serve any informative purpose
- Changed the scope of Resource's ResourceKey to protected, so subclasses could access the resource name

---ZOMBIE RESOURCES---
This change allows for the creation of "zombie" resources when loading fails. Such resources won't cause problems or error out when invoked by the engine, but they won't do anything. Resource constructors now only initialize the resource name. All actual resource loading from file happens in a separate "load" method. Each Resource subclass has been updated accordingly. The ResourceLoader now creates resources and loads in resource data in two separate steps. When a load fails, the uninitialized resource is still returned and expected to act as a stub in order to allow the game to continue. This uninitialized resource is not placed in the resource map, because we want the ResourceLoader to retry loading the data on later requests until the load succeeds. This solution creates a memory leak because the zombie objects are currently entirely unmanaged. Memory leaks are strictly better than segmentation faults, but this problem will need to be addressed in the next check-in.

Next steps:
Fix the memory leak. Then fix up Tileset, Region and Spritesheet loading to also create harmless zombie state resources. Finally, with resource failure handled, we can move to proper script handling when there is an error either loading or running a script.

git-svn-id: http://robothaus.org/repos/ed@59 5033be24-7f03-4d61-c009-87962669689c