Few changes in this bundle of code:
We now refer to NPCs via string names rather than making each NPC a global variable with its own name. This makes it easier to keep track of an NPC, even if its script has problems executing.

Error handling for scripts is nicer now; if any Thread throws an exception, then the Scheduler will catch the Exception and take the thread out of rotation. Lua script failures in global (non-NPC) scripts throw exceptions on failure now, indicating that they should be removed. NPC scripts that are not working will fail on creation (since all their code and functions are loaded on creation). Thus, if there is a failure, the functions that were successfully loaded will be noted, and the ones that weren't loaded won't be called. To do this, we have an array of booleans indicating whether or not each function exists for a given NPC. This will always be false if a function is coded incorrectly or does not exist. This also replaces our need for an empty function to call if the NPC doesn't have a certain function.

So now if a script fails, we won't segfault or freeze or even flood debug output with the "resume non-suspended coroutine" Lua error message. Instead, the problematic NPC won't run its problematic functions (and any functions below it in the Lua script file), and you can read the debug logs to find out why.

The "move" function for NPCs now gets piped through the Tile Engine, which uses the NPC list to look up the name of the NPC and then calls that NPC's move function. Other NPC functions may be coded this way for the next while.

Spritesheet error handling is also included; if the Spritesheet is not properly constructed, then it will not be drawn to screen. So instead of a segmentation fault, an NPC will just disappear, and you can read the debug logs to determine why.

Updated the to-do list, as usual. I'm almost done with "nice" error handling, but I'd like to take a break and continue to get the game features going. The code is modular enough that I can revisit the "nice" handling issue at any time.

git-svn-id: http://robothaus.org/repos/ed@65 5033be24-7f03-4d61-c009-87962669689c