Anyway, before doing that, some small changes: renamed TitleScreenListBox to ListBox, and created an action on the main menu that accesses the Menu prototype with some mock data.
I realized that to start working on the menu, it'd be handy to have a (relatively) fleshed out character model. And then I also realized that although I use XML, I don't really have any deep tree structures; just single elements with attributes. So I'm going to experiment with SQLite, see how easy it is to work with compared to XML files.

git-svn-id: http://robothaus.org/repos/ed@80 5033be24-7f03-4d61-c009-87962669689c