Save data changes: After struggling with the SQLite decision, I realized that the only reason the save data looked relational is because I set it up in a relational way, rather than taking advantage of XML's strong point (tree structure). Instead of using a primary key on characters (their names) and linking it to a master character list, I'm now keeping lists of Characters, with all their stats, in their respective locations in the player's group (in the party or in the reserve, right now).
Menu prototype:
Added menu prototype panels, which can and will slowly be fleshed out. Menu populates tabs with the player save data given, and passes it to the panels, where they can draw the relevant data out of the player data. The Menu maintains a stack to determine which panels have been shown and what to show when the player cancels out of the current panel. When there are no active panels left, the player has cancelled out of the Menu and back to the state from which they accessed the Menu.

This was a good return to the visual stuff after a lot of slogging around with map stuff. I love this project. <3

git-svn-id: http://robothaus.org/repos/ed@81 5033be24-7f03-4d61-c009-87962669689c