Switched the savegame format to JSON because I like JSON more for this kind of data serialization. Also JsonCpp is much more intuitive than tinyXML for this kind of thing, in my opinion.
Added item metadata so that item schemas can be loaded from file. They contain all the relevant item data (or will, at least), so the goal here is to employ the Flyweight design pattern to items instead of recreating the same item structure over and over again for no reason.

The item metadata is loaded into a new singleton that loads the items.edb in the new metadata folder, which will also eventually contain monster data and other schematic metadata for the game.

I'm keeping the XML version of the savegame for now just in case there are discrepancies between it and the newly created JSON version, but I will probably remove the file on the next commit.

Also, I added a headers group to the CMakeLists.txt so that headers will also be loaded into the project files that are created. I will likely divide the compilation units into more source groups so that XCode won't just list all the files in one project folder (which makes it very hard to navigate between files!).

Finally, got rid of a couple of compile-time warnings that really should be treated as errors; use of std::string as a parameter to a format string. Since std::string is a non-POD type, DEBUG statements that use one for a %s parameter will cause a crash. A conversion using a c_str() call is necessary here.



git-svn-id: http://robothaus.org/repos/ed@94 5033be24-7f03-4d61-c009-87962669689c