Refactoring menu, fixing hacks and warnings.
Refactored the ListBox object to manage less things and not push events
to SDL. It seemed like a smart idea at the time, to just push messages
from the list boxes and let the main loop manage them, but it's an
awkward design in terms of tracing code and also for readability. Using
listeners and injecting them as needed made a lot more sense, and keeps
the logic in the game state rather than in the widget code.

Also fixed the need for a public list box in the menu shell, which was a
hack from the start. Removed about 700 GCC warnings as well. And hey!
This is my first git commit for this project. That's pretty neat. <3

git-svn-id: http://robothaus.org/repos/ed@100 5033be24-7f03-4d61-c009-87962669689c