Modal Yes/No dialog, game save fixes, gcn Window extension, Data Menu <3
A modal state called "ConfirmState" was introduced to allow for creation
of yes/no dialogs that force a user to respond. To get this modal focus
working, a fix was needed to the way we detach containers when switching
game states.

When Gui::setTop is called, the old top container is removed, and the
focus handler is removed from the old container as well. However, mouse
listeners and other action listeners are still holding onto the
now-detached container, and so Guichan hits an exception when we request
focus and have to touch these listeners.
The solution is to have a top-level container, and add the game state
GUIs to the top-level container when a state is created. Instead of
removing the container, we set it to be invisible. In this way, there
are no GUI elements that the Gui object isn't attached to.

The serialization of Quest objects for PlayerData saving was done
incorrectly. Sorry about that. Quest serialization is now fixed up.

Extending the gcn::Window class and made an edwt::Window to fix a modal
focus issue where any mouse dragging above the window would move the
window.

Finally, the Data Menu and Data Pane were fleshed out, and now the user
can save over other games (with confirmation of course, hence the
Confirm dialog above!). I'm thrilled to have yet another piece finished
(though the GUI will need to be laid out and styled, of course).

git-svn-id: http://robothaus.org/repos/ed@101 5033be24-7f03-4d61-c009-87962669689c