Pathfinding debug tool and bug fix
Fixed the NPC pathfinding to find a path only right before an NPC is
about to move (since move orders can be queued up and may all share the
same source tile otherwise).

Added a debugging mechanism: the ability to draw NPC tasks. In this
case, MoveTasks now draw a red line indicating the remaining path that
the NPC will traverse. This will make pathfinding glitches easier to
capture in the future.