A* pathfinding (!), plus new debug tool and updated Pathfinder API
Finally! NPCs can now find paths using A* search, which takes into
account dynamically added obstacles and moving entities such as the
player or other NPCs.

Collision detection still doesn't exist, and will be part of the next
commit. But the groundwork is in place to add that in correctly, and
the Pathfinder functions have been updated to more clearly express how
entities should call it to check for collisions.

Also added a new debug macro that makes the currently loaded Map draw
the Pathfinder's collision map, to more easily debug pathfinding and
collision detection issues.

Augmented Point2D to have get*AdjacentPoints() functions, which allow
for an easy way to get a tile's neighbors without rewriting the same
eight if-statements over and over again. I expect that it will be
useful in a few places in the Tile Engine component.

Next up is collision detection, as well as some review and cleanup of
the Pathfinder class, which has become somewhat complex and
algorithms-heavy and still lacks correct header documentation.