Fixes to pathfinding, NPCs always occupy space
The NPC Move order uses incorrect locations to signal movement ending
when a chain of paths are followed. This caused NPCs to leave a trail
of false obstructions that would permanently block movement for other
entities for no reason.

The Tile Engine did not check whether or not NPCs can stand in certain
locations before spawning them. At this point, NPCs do not spawn at all
if they cannot occupy their assigned space (if they are spawned into a
wall or on top of another NPC, for instance). NPCs now always occupy
space from the beginning of their lifetime until the end.

At this point, dynamic pathfinding works well! :D
Some code cleanup is required in NPC_MoveOrder though, and Pathfinder
needs to be moved up to the TileEngine level as mentioned in the last
commit.