Small bug fixes, new lookAtPlayer script function.
NPC now turns to player and says Hello! :D

Fixed up segfault when an instruction in the NPC's activate function is
blocked by the scheduler.

Fixed the way scripts continue after waiting for a dialogue line to
complete. They now continue after the line is cleared from the dialogue
box, rather than when they are displayed on screen.

Fixed the height adjustment for the dialogue box widget.

Finished off Stand function, and used it while adding a new
lookAtPlayer function for NPCs to turn towards the player.

Also fixed up copyright notice for new files added in the last commit.