Updated Spritesheet reader to ingest TagSprite JSON
The Spritesheet class now reads in an EDS file that contains JSON
instead of the current format. There were a few reasons for this:

1) Easier to catch errors by imposing a structure on the file.
2) Easier to change or add fields later, and order doesn't matter for
different data fields.
3) Parsing code is much cleaner since the syntactic checks are handled
by JsonCpp.

In addition to this, updated npc1.png to have some padding between all
its frames (so TagSprite could work on it properly), and used TagSprite
to produce the new npc1.eds file.

From here, it should hopefully be much easier to make some much needed
changes to the spritesheett data.