Updated sprite selection logic to include character direction
Updated npc1.eds animations to remove standing frame (looked a bit
jerky with this frame included).

Updated Animation object to hold its own name. The Sprite class
performed an equality comparison between Animations, but since this
compared pointers and new Animations were always created in a
Spritesheet::getAnimation call, this was bugged. Even if it had worked,
there was a small memory leak that would have been caused by the
ensuing code.

Sprite object keeps track of the current frame/animation name and the
direction in order to save a call to Spritesheet::getFrame or
Spritesheet::getAnimation in case of a repeated call.

NPC and PlayerCharacter now initialize currDirection. It's very scary
to me that enum values can be uninitialized (and can therefore be
out-of-bounds values) but they're just slightly restricted ints so I
understand the reasoning.

Removed lots of hacky logic for changing sprite direction in MoveOrder
and StandOrder. Sprite now takes direction as an argument and tries to
find the frame with the appropriate direction, so that calling classes
don't need to know the directed sprite names in advance (will be
impossible once content grows).