PlayerCharacter is now properly blocked by walls, obstacles and NPCs.
Player collision detection is done by moving in half-tile steps on the collision map and trying to occupy each spot along the way (and freeing the one behind it) until the full movement completes or an obstruction/map edge is encountered.

This method is less general/elegant than what I had initially planned on, but my more general algorithm was much less readable and prone to error. This, on the other hand, does the job really well in way
less code, and it performs well enough for now since most frames don't trigger more than two attempts to occupy tiles. Hopefully it will continue to perform well, but if not then we can revisit the function at a later time.