Added width/height considerations to pathfinding, more granular tiles, bug fixes.
Made movement tiles in the Pathfinder 16 pixels instead of 32 (and in
doing so, found and fixed an array overflow that was waiting to
happen). The area that characters tend to take up has been
significantly trimmed down as a result, and characters no longer have a
huge amount of space between them when they move up against each other.

Refactored Pathfinding to trim the debug log flooding that happened
whenever the player tried to move. Added canOccupyArea to test if an
area could be acquired without actually acquiring it.

Added width and height as parameters to the A* pathfinding algorithm.
This prevents the pathfinding algorithm from finding a path that the
caller cannot actually fit into.

Finally, fixed a potential infinite loop when the player blocks the
path of an NPC.