Fixing NPC pathfinding bugs (diagonal pathfinding, NPC livelock)
NPCs would, when obstructed, reroute using A*. However, the A*
algorithm didn't do a thorough enough check to determine whether or not
an NPC could indeed move diagonally.

Additional checks were added to ensure that the lateral spaces
surrounding a diagonally adjacent neighbor were free as well.
Otherwise, the NPC would calculate a diagonal path that skimmed an
obstacle, but would be unable to actually circumvent the obstacle, thus
locking up.

Another point of contention was that if an NPC took up multiple tiles,
A* would wrongly discard all but one tile occupied by the NPC. The
canOccupyArea function is now used in the A* node evaluation to keep
things consistent and prevent false discards.

With that, I think NPC and player movement are pretty tight. I say it's
time to tie the TileEngine and the player data together so we can get
some fetch quests "up in". :)