Added functions to add/remove player from map. Started integrating player data into game.
Added three functions to the Lua scripts: setPlayerLocation, showPlayer and removePlayer. Also changed addNPC function to take parameters in pixels rather than in tiles.

From now on, when a new region or map is set, the player entity is not on the map until a script invokes showPlayer and gives the player a legitimate location.

Also added the player data to the tile engine and script engine so that the game can start interacting with the player data.

Also fixed some bugs and cleaned up the Script Engine code a bit.