Added C++/Lua binding using LuaWrapper
Originally, the thought was that I could defer object binding and treat
it as a nice-to-have rather than an important feature. However, all the
glue that's needed for each new script function is making the code
messy (particularly ScriptFunctions.cpp).

Using the LuaWrapper library, I can bind new script functions directly
to the objects they affect. As a result, the added indirection from Lua
to the ScriptEngine class to the TileEngine class is not necessary.

Currently checking in preliminary work, exposing a few PlayerCharacter
functions. Converting more functions to this style in the next
check-ins. :)