Added NPC bindings in Lua
Added NPC bindings in Lua (along with proper pushing of the NPC
parameter in idle or activate functions).

Included in this change are the functions "addNPC" (behaves the same as
before, but is now bound to the TileEngine object), and "move" (acts on
an NPC).

Added a global to the script engine, called "player", which references
the PlayerCharacter object held by the tile engine.

Altered all the scripts to be compatible with this new syntax and I
must say… it looks pretty!