Fixing a Lua VM corruption bug.
The return value for the random function was being pushed onto the main VM instead of the calling NPC's stack. This led to the NPC pulling bad values off the stack and the main stack getting random numbers pushed on it. The result was corruption of the Lua VM; the game eventually segfaulted whenever the scripts invoked the "random" function repeatedly.