Binding inventory to Lua scripts
Added bindings for the player's inventory so that Lua scripts can query
and modify it.

To facilitate this, an Inventory class was created to encapsulate an
ItemList and expose functionality to scripts or the engine. After this,
the Inventory class was bound to Lua, and a global named "inventory"
was created and linked to the loaded PlayerData inventory field.

The scripts were updated to demo some of the new functionality.