Added exits/entrances to maps
Using the Message Pipe work from the last commit, added a new
MapExitMessage that is triggered by the EntityGrid. Also removed
TriggerDetector and moved trigger detection code into EntityGrid.

The Map class is augmented with data about exit zones (MapExit objects)
that lead to other maps, as well as entry points from maps that exit
into it.

Registered the Tile Engine to listen for and act upon exit messages by
changing the map and then putting the player at the correct entry point
on the new map, based on the name of the old map.

Added a second example map to test exits.

Added message pipe to Actor class so that it could send
ActorMoveMessages to the message pipe.

Known issues from this change:
1) Slight lag spike when the player crosses an exit and changes maps.
2) NPCs from previous map remain along with their scripts. Generally,
cleanup of the previous map at the Tile Engine level is not done.