Created class to manage OpenGL textures, fixed alpha testing
Created the Texture object to encapsulate an OpenGL texture. This
object is more in line with RAII, since it creates a texture on
construction and deletes it on destruction. Later on, when we start
using C++11, we can wrap this in a shared pointer and make some of our
image resources more exception-safe, something I haven't been paying
nearly enough attention to unfortunately.

While doing this, I stepped through the texture creation code and
discovered an issue. We blit image textures to an RGBA surface in order
to convert all images to RGBA before sending them to the GPU. But this
blotting process is dropping the alpha channel for some reason. Instead
of this approach, we will determine the format of the image and send
that to the GPU without an intermediary surface. After this, alpha
testing for sprites started working again. No more black boxes around
the characters! :D