Fix issue where NPCs aren't cleaned up on map change
Properly cleaning up the list of NPCs and deleting them when the Tile
Engine instance is destroyed or when the region/map changes.

Now, when the player moves to a different map, the NPCs don't stay on
the screen and keep walking around. When the player returns to the
first map, the NPCs are back in their beginning states.