Initial commit for foreground layer support.
Created a Layer class that does the rendering work inside a Map. Map objects now load and create multiple background/foreground layers and render them in order of arrangement in the map editor.

Also made a change to much of the OpenGL code to use glPushAttrib/glPopAttrib instead of repeated glGet calls.

Also added a new tileset with houses and trees. Unfortunately, DanC's Hard Vacuum tileset, though really nice, doesn't have enough multi-tile objects to really test the foreground layer support, so this tileset had to be put away for now.

The layer work still has a lot of bugs to work through, which will be fixed in the next commit. The main issue is proper back-to-front rendering order for the layers and the sprites.