Finalizing layer support in tile engine
The tile engine, instead of just drawing the background, then sprites,
then foreground, now draws the background, sprites and foreground
row-by-row to ensure proper drawing order. All the background is drawn,
then for each row, the sprites and then the foreground on that row are
drawn.

In order to get the layers working properly, a concept of "height"
needed to be added to the layers. Based on the "heightOffset" property,
optional in a layer, the tiles in a layer are offset by heightOffset
rows, and then drawn heightOffset tiles above their location. In other
words, they are drawn where they were placed in the Tiled Map Editor,
but logically they are located heightOffset rows down.

The problem with the approach used is that Layer needs to actually
translate all of its tile data down by heightOffset rows on load, so
that the map renders in EDEn the same way that it does in Tiled.
Ideally, Tiled would support a layer-based drawing offset. That way, a
user could place a tile in the right logical location, see it being
drawn in the right place in Tiled, and see it being drawn in EDEn
without translation. Tiled is tracking a feature request for
layer-based offsets in issue 4. Once this feature is added,
heightOffset needs to be removed, and the y-offset feature must be
honoured by EDEn instead.

The result of all this work is that layers are now drawing correctly
and not clipping the player and NPC sprites in weird ways.

Also changed:
- Entity Grid debug output was trimmed
- Crash on quit that was introduced in commit
4d9ba875e5465cfeacd5802f1936db29deaefa0e was fixed