Initial work for transitions between states. Adds the ScreenTexture object, which creates a frame buffer and binds rendering operations to it so that drawing can be captured in a texture.
The first experimental transition is a simple fade to black. This isn't too different visually than what we already have, except that the previous version was just blitting over whatever was on the screen enough times to make the whole thing black. This version doesn't block the UI thread during the fade, and it's more precise.

There is still some work to be done around making the Execution Stack handle transitions more cleanly, because currently it doesn't support a game state pushing a state and then finishing in the same frame (causes the newly pushed state to get deleted instead of the transition state). As well, there is a really weird bit in the FadeState.cpp draw method that pinches in the screen capture by 0.5f on each edge. For some reason, the render looks weird without this, but I would like to find out why this is happening before merging this into the main dev branch.