Improvements to transition framework
Making ExecutionStack aware of transition states so that it can clean
up a transition state before the next state is activated. Prior to this
change, transition states had to remain on the stack and execute a
completion frame when the state above it in the stack was completed.

Aside from this, some small cleanup, and removal of newState parameter
in SpinState and FadeState, since these two states only act on the old
state.