Pulling common transition code up to TransitionState superclass
Created a TransitionState class that will calculate progress through a
transition and defer drawing to subclass transitions. This superclass
will also be responsible for initializing the ScreenTextures needed by
each transition.

Made BlendState, SpinState and FadeState into subclasses of
TransitionState. None of them have a step function anymore, and their
constructors are empty, because all that functionality was the same
across the three of them and thus was pulled up to TransitionState
instead.