Adding CharacterRoster, plus bindings to Lua for Characters
Moved the character management in PlayerData into a new CharacterRoster
class. This class manages the party, the current lead character, as
well as the list of all characters encountered by the player.

Updated references to PlayerData's character management to go through
the roster instead (mostly updates to Menu).

Added Lua bindings for Character and CharacterRoster. The
CharacterRoster currently just allows for adding a Character to the
party. The Character binding allows Lua scripts to query the name of
the character. The next updates will flesh out these bindings and test
them with updates to the existing Lua scripts.