Added Lua hook for creating characters
Added a method to the CharacterRoster object in Lua to create a new
character from an archetype. Added a character to the player's party in
the chapter1 script for testing this capability.

Also added sprite sheet data to the character archetype. This data is
not yet used, but will be used by the PlayerActor soon.

Fixed a small bug in CharacterRoster where the character lookup in
addToParty indexed on the character name instead of the character ID.
This was causing a character to appear in the character list multiple
times, and subsequently get deleted more than once.