Bound PlayerCharacter to sprite sheet of the Player's Party Leader
A new RosterUpdateMessage was added to the message pipe to process
changes to the roster, party and party leader. The PlayerCharacter used
by the Tile Engine now binds to this property to keep itself in sync
with the current party leader. When the leader changes, the
PlayerCharacter will use the new leader's sprite sheet immediately.

CharacterRoster now fires a RosterUpdateMessage whenever it changes.

PlayerData (and the CharacterRoster) now bind to a Message Pipe. Since
the PlayerData is passed between states, the message pipe is re-bound
whenever the state changes. In this case, the Tile Engine clears the
message pipe property of PlayerData whenever a new state is pushed, and
re-binds the message pipe on its activate event.