Adding extension manager to fix build break on Windows
The Windows build broke due to the new extensions used by ScreenTexture
for FBO support. In order to make sure that the game compiles and links
correctly, a dynamic approach is now used.

The OpenGLExtensions class, instantiated in GraphicsUtil, will load the
extensions using SDL_GL_GetProcAddress. It exposes these loaded extensions
and flags indicating whether or not the extensions were successfully
loaded.

The ScreenTexture class checks if FBO support is available before
attempting to use it.