Code cleanup - Removing old code and enforcing const-correctness
Went through all of the older source files in the codebase and ensured
that methods that shouldn't change state are declared const.

Added a TODO around making EquipData::getSlots a const function, since
modification of the data really should be done using the EquipData
class itself.

Removed SpriteFrame because it was literally just a rectangle, and we
have shapes::Rectangle for that purpose.