Redefining debug switches to be checked at runtime
Redefining debug switches so that they are always some number and
always defined. Wrapping checks in if-statements rather than #ifdef
statements. This is being done because the #ifdef checks ensured that
debug code wasn't always compiled, meaning that build errors in those
blocks were only found when the debug switches were turned on. With
this modification, debug blocks are always compiled, even when they
aren't used, so broken code is caught immediately.