Adding support in Rocket for SDL_Image's supported types
Using the Texture class in the Rocket render interface instead of the
tutorial's rendering code. The result is that more image formats are
supported, and the Texture class can be used for all texturing.

The Texture class now supports loading texture data from memory.

Also updated the copyright text and documentation in the interface files. As the ShellRenderInterfaceOpenGL class is now very heavily modified and somewhat EDEn-specific, updated the documentation and class name accordingly. The files will be renamed to EdenRocketRenderInterface to match the new class name in the next commit.

Overriding the regular Rocket logging to use DebugUtils, so we don't have two different debug logging systems working independently.

Removed the title screen's Guichan interface. The next step is to pipe
input into Rocket and receive click events to get back to parity with
the previous title screen implementation.